refactor(server): break up monolithic llm.ts into focused modules (#296)
Split the 1122-line God file into focused modules:
- llm-types.ts: shared StreamChunk/WorkflowStep types (avoids circular deps)
- tool-executor.ts: shared tool-call execution utility
- backends/anthropic.ts: Anthropic API backend with tool-call loop
- backends/cli.ts: Claude CLI subprocess backend
- backends/openai.ts: OpenAI-compatible backend (Ollama, llama.cpp, OpenAI)
- llm.ts: slim orchestrator (~220 lines) that delegates to backends

All existing public exports remain available via re-exports.
